pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-audit"
version = "0.1.0"
description = "Estimate costs and lint agent workflow YAML files"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "AreteDriver"}]
keywords = ["ai", "agents", "workflow", "cost", "linter", "claude", "openai", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent-audit = "agent_audit.cli:app"

[project.urls]
Homepage = "https://github.com/AreteDriver/agent-audit"
Repository = "https://github.com/AreteDriver/agent-audit"
Issues = "https://github.com/AreteDriver/agent-audit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agent_audit = ["data/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/agent_audit/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
